IAmViral/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.8.0
numpy==1.21.2
Pillow==8.3.1
requests==2.26.0
python-dotenv==0.19.0
tweepy==4.5.0
pytrends==4.8.0
scikit-learn==0.24.2
pandas==1.3.3
TikTokApi==5.2.0
instaloader==4.10.0 